requirements.txt
datefinder==0.7.3
easyocr==1.7.1
fastapi==0.103.1
numpy==1.25.2
opencv_python_headless==4.8.0.76
Pillow==10.0.0
uvicorn==0.23.2
python-dateutil==2.8.2
python-multipart==0.0.6
regex==2023.8.8
imutils==0.5.4